feat:double qoute に変更した

+++ functions/.eslintrc.js
@@ -26,7 +26,7 @@ module.exports = {
     "import",
   ],
   rules: {
-    "quotes": ["error", "single"],
+    "quotes": ["error", "double"],
     "import/no-unresolved": 0,
     "indent": ["error", 2],
   },
